% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundid_adjust.R
\name{boundid_adjust}
\alias{boundid_adjust}
\title{Prepare weighting or trimming parameters before the estimation step}
\usage{
boundid_adjust(
  Data,
  X_var,
  threshold,
  floor = TRUE,
  treatment,
  time,
  group2_mean = NULL,
  ID,
  ATU = F,
  ATT = F,
  S = NULL,
  cut = FALSE,
  panel = F,
  orig_ID
)
}
\arguments{
\item{Data}{An R dataframe.}

\item{X_var}{The outcome variable.}

\item{threshold}{The natural boundary of the outcome.}

\item{floor}{A logical statement. Should be set to TRUE if the boundary is a floor, or to FALSE if the boundary is a ceiling. Default is TRUE.}

\item{treatment}{The treatment variable.}

\item{time}{The time variable.}

\item{group2_mean}{A numeric value indicating the pre-treatment mean of the treatment group.}

\item{ID}{The unit ID.}

\item{ATU}{A logical statement. TRUE if the ATU is the target estimand.}

\item{ATT}{A logical statement. TRUE if the ATT is the target estimand.}

\item{S}{= A numeric value indicating the counterfactual fall. If NULL, the function automatically computes the counterfactual fall of the control group.}

\item{cut}{A logical statement. If TRUE, then the cutting/trimming approach will be used. If FALSE, then the weighting approach will be used.}

\item{panel}{A logical statement. If TRUE, then the function generates weights for the whole panel. If FALSE, then the function generates weights for two time periods.}
}
\value{
The output of this function is a vector containing the newly generated weights to be included in the estimation step to counter-act boundary bias.
}
\description{
Prepare weighting or trimming parameters before the estimation step
}
\examples{
new_ATT_weights <- boundid_adjust(simulated_data, "X", 0, floor = T,
treatment = simulated_data$treated, time = simulated_data$time,
ID = simulated_data$ID, ATT = T, cut = F, panel = F,orig_ID = simulated_data$ID)
}
