% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundid_test.R
\name{boundid_test}
\alias{boundid_test}
\title{Evaluate (the extent of) boundary bias in a difference-in-differences setup}
\usage{
boundid_test(
  Data,
  X_var,
  threshold,
  floor = TRUE,
  treatment,
  time,
  ID,
  first_period,
  S = NULL,
  more_info = F,
  stag = FALSE
)
}
\arguments{
\item{Data}{An R dataframe.}

\item{X_var}{The outcome variable.}

\item{threshold}{The natural boundary of the outcome.}

\item{floor}{A logical statement. Should be set to TRUE if the boundary is a floor, or to FALSE if the boundary is a ceiling. Default is TRUE.}

\item{treatment}{The treatment variable.}

\item{time}{The time variable.}

\item{ID}{The unit ID.}

\item{first_period}{A numeric value indicating the first time period in which the treatment is administered.}

\item{S}{= A numeric value indicating the counterfactual fall. If NULL, the function automatically computes the counterfactual fall of the control group.}

\item{more_info}{A logical statement. If TRUE, then additional information is included in the output, such as the distance of the furthest unit above or below the boundary (or the distance of the 1st percentile)}

\item{stag}{A logical statement. Should be TRUE if the design is staggered or FALSE for a basic 2x2 DiD.}

\item{cut}{A logical statement. If TRUE, then the cutting/trimming approach will be used. If FALSE, then the weighting approach will be used.}
}
\value{
The output of this function is a vector containing the newly generated weights to be included in the estimation step to counter-act boundary bias.
}
\description{
Evaluate (the extent of) boundary bias in a difference-in-differences setup
}
\examples{
boundid_test(simulated_data, "X", 0, floor = T, simulated_data$treated, simulated_data$time, more_info = F, ID = "ID")
}
